MidiConvert/cbo/requirements.txt
midi2audio
music21
pydub
pygame
midiutil
pillow